+++ src/packages.h
@@ -5,6 +5,7 @@
 #include <stdbool.h>
 
 #include "ast.h"
+#include "stdlib/datatypes.h"
 
 Text_t get_package_name(Path_t lib_dir);
-OptionalPath_t find_installed_package(ast_t *use);
+OptionalPath_t find_installed_package(Table_t *build_info, ast_t *use);