+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## 2026-04-27
+
+- Add build metadata and version information, which can be retrieved with `tomo -b`.
+- Change interface for `List.binary_search()` to be more flexible.
+
 ## 2026-03-16
 
 - Added `Path.copy_to(src, dest)`